LayoutTests/fast/css/css-value-wrapper-sameness.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<style>
.a { width: 0px; height: 0px; }
.a { width: 100px; height: 100px; }
.a { width: 1000px; height: 1000px; }
.a { width: 50%; height: 50%; }
.a { color: red; background-color: red; }
.a { color: rgb(0,0,0); background-color: rgb(0,0,0); }
.a { color: rgb(255,255,255); background-color: rgb(255,255,255); }
.a { color: rgba(0,0,0,255); background-color: rgba(0,0,0,255); }
.a { color: rgba(0,0,0,0); background-color: rgba(0,0,0,0); }
.a { color: inherit; background-color: inherit; }
.a { color: initial; background-color: initial; }
</style>
<script>

description("This test documents the sameness characteristics of CSSValue wrappers (unshared, unique objects.)");

rules = document.getElementsByTagName('style')[1].sheet.cssRules;
for (i = 0; i < rules.length; ++i) {
    style = rules[i].style;
    value0 = style.getPropertyCSSValue(style.item(0));
    value1 = style.getPropertyCSSValue(style.item(1));
    shouldBeFalse("value0 === null");
    shouldBeFalse("value1 === null");
    shouldBeFalse("style.getPropertyCSSValue(style.item(0)) == style.getPropertyCSSValue(style.item(0))");
    shouldBeFalse("value0 === value1");
    shouldBeFalse("value0 == value1");
    shouldBeTrue("value0.cssText == value1.cssText");
}

</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
