LayoutTests/fast/css/css-value-wrapper-sameness-expected.txt

This test documents the sameness characteristics of CSSValue wrappers (unshared, unique objects.)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS value0 === null is false
PASS value1 === null is false
PASS style.getPropertyCSSValue(style.item(0)) == style.getPropertyCSSValue(style.item(0)) is false
PASS value0 === value1 is false
PASS value0 == value1 is false
PASS value0.cssText == value1.cssText is true
PASS value0 === null is false
PASS value1 === null is false
PASS style.getPropertyCSSValue(style.item(0)) == style.getPropertyCSSValue(style.item(0)) is false
PASS value0 === value1 is false
PASS value0 == value1 is false
PASS value0.cssText == value1.cssText is true
PASS value0 === null is false
PASS value1 === null is false
PASS style.getPropertyCSSValue(style.item(0)) == style.getPropertyCSSValue(style.item(0)) is false
PASS value0 === value1 is false
PASS value0 == value1 is false
PASS value0.cssText == value1.cssText is true
PASS value0 === null is false
PASS value1 === null is false
PASS style.getPropertyCSSValue(style.item(0)) == style.getPropertyCSSValue(style.item(0)) is false
PASS value0 === value1 is false
PASS value0 == value1 is false
PASS value0.cssText == value1.cssText is true
PASS value0 === null is false
PASS value1 === null is false
PASS style.getPropertyCSSValue(style.item(0)) == style.getPropertyCSSValue(style.item(0)) is false
PASS value0 === value1 is false
PASS value0 == value1 is false
PASS value0.cssText == value1.cssText is true
PASS value0 === null is false
PASS value1 === null is false
PASS style.getPropertyCSSValue(style.item(0)) == style.getPropertyCSSValue(style.item(0)) is false
PASS value0 === value1 is false
PASS value0 == value1 is false
PASS value0.cssText == value1.cssText is true
PASS value0 === null is false
PASS value1 === null is false
PASS style.getPropertyCSSValue(style.item(0)) == style.getPropertyCSSValue(style.item(0)) is false
PASS value0 === value1 is false
PASS value0 == value1 is false
PASS value0.cssText == value1.cssText is true
PASS value0 === null is false
PASS value1 === null is false
PASS style.getPropertyCSSValue(style.item(0)) == style.getPropertyCSSValue(style.item(0)) is false
PASS value0 === value1 is false
PASS value0 == value1 is false
PASS value0.cssText == value1.cssText is true
PASS value0 === null is false
PASS value1 === null is false
PASS style.getPropertyCSSValue(style.item(0)) == style.getPropertyCSSValue(style.item(0)) is false
PASS value0 === value1 is false
PASS value0 == value1 is false
PASS value0.cssText == value1.cssText is true
PASS value0 === null is false
PASS value1 === null is false
PASS style.getPropertyCSSValue(style.item(0)) == style.getPropertyCSSValue(style.item(0)) is false
PASS value0 === value1 is false
PASS value0 == value1 is false
PASS value0.cssText == value1.cssText is true
PASS value0 === null is false
PASS value1 === null is false
PASS style.getPropertyCSSValue(style.item(0)) == style.getPropertyCSSValue(style.item(0)) is false
PASS value0 === value1 is false
PASS value0 == value1 is false
PASS value0.cssText == value1.cssText is true
PASS successfullyParsed is true

TEST COMPLETE
